INS-654: bootstrapcertificate moved to certificate component as is
